move excess processing to data functions
