+++ dashboard-ui/src/components/AggregateResults/DataFunctions.js
@@ -18,6 +18,18 @@ const SIM_MAP = {
     "dryrun-soartech-eval-vol3": 3
 }
 
+const adept_dre_names = {
+    '1': 'DryRunEval-MJ2-eval',
+    '2': 'DryRunEval-MJ4-eval',
+    '3': 'DryRunEval-MJ5-eval'
+}
+
+const st_dre_names = {
+    '1': 'QOL-VOL-1',
+    '2': 'QOL-VOL-2',
+    '3': 'QOL-VOL-3'
+}
+
 const MED_ROLE_MAP = {
     "M-3 Medical student": 1,
     "M-4 medical student": 1,
@@ -653,6 +665,23 @@ function populateHumanDataRow(rowObject, version) {
     return returnObj;
 }
 
+function getGroupKey(row, selectedEval) {
+    if (selectedEval === 3) {
+        return row.SimEnv;
+    } else if (selectedEval === 4) {
+        const adeptName = adept_dre_names[row.ADEPT_Scenario] || row.ADEPT_Scenario;
+        const stName = st_dre_names[row.ST_Scenario] || row.ST_Scenario;
+        return `${adeptName}_${stName}`;
+    }
+}
+
+function formatCellData(data) {
+    if (typeof data === 'object' && data !== null) {
+        return JSON.stringify(data);
+    }
+    return data;
+}
+
 function populateDataSet(data) {
     let simAlign = null;
     if (data.getAllSimAlignmentByEval) {
@@ -1171,4 +1200,4 @@ function getChartData(data) {
 }
 
 
-export { populateDataSet, getAggregatedData, getChartData, isDefined };
+export { populateDataSet, getAggregatedData, getChartData, isDefined, getGroupKey, formatCellData };
